+++ example/js/api/ICU4XFixedDecimalGroupingStrategy.mjs
@@ -1,0 +1,15 @@
+import wasm from "../wasm.mjs"
+import * as diplomatRuntime from "./diplomat-runtime.mjs"
+
+export const ICU4XFixedDecimalGroupingStrategy_js_to_rust = {
+  "Auto": 0,
+  "Never": 1,
+  "Always": 2,
+  "Min2": 3,
+};
+export const ICU4XFixedDecimalGroupingStrategy_rust_to_js = {
+  0: "Auto",
+  1: "Never",
+  2: "Always",
+  3: "Min2",
+};
