+++ example/js/api/ICU4XFixedDecimalSignDisplay.mjs
@@ -1,0 +1,17 @@
+import wasm from "../wasm.mjs"
+import * as diplomatRuntime from "./diplomat-runtime.mjs"
+
+export const ICU4XFixedDecimalSignDisplay_js_to_rust = {
+  "Auto": 0,
+  "Never": 1,
+  "Always": 2,
+  "ExceptZero": 3,
+  "Negative": 4,
+};
+export const ICU4XFixedDecimalSignDisplay_rust_to_js = {
+  0: "Auto",
+  1: "Never",
+  2: "Always",
+  3: "ExceptZero",
+  4: "Negative",
+};
