+++ feature_tests/js/api/ErrorEnum.mjs
@@ -1,0 +1,11 @@
+import wasm from "../wasm.mjs"
+import * as diplomatRuntime from "./diplomat-runtime.mjs"
+
+export const ErrorEnum_js_to_rust = {
+  "Foo": 0,
+  "Bar": 1,
+};
+export const ErrorEnum_rust_to_js = {
+  0: "Foo",
+  1: "Bar",
+};
